Refactor API service layer for consistency
This commit refactors the frontend API handling to use a single,
centralized service layer built on top of the existing axios client.

Previously, API calls were scattered across multiple files with
duplicated logic, hardcoded URLs, and inconsistent error handling.
This made the code difficult to maintain and debug.

Changes in this commit:
- Created a new `src/services` directory to house all API service files.
- Consolidated API calls for AI, search, and explore into dedicated
  service files (`ai.service.ts`, `search.service.ts`, etc.).
- All services now use the centralized `apiClient` from `src/lib/axios.ts`,
  ensuring consistent base URL, error handling, and authentication.
- Refactored `useSmartAI`, `ContextualSearchResults`, and `SmartSearchBar`
  to use the new services.
- Replaced a mock search implementation in `contextualSearchAPI.ts` with
  actual API calls.
- Removed the old, redundant API utility files.
